+++ crates/mirror-mirror-macros/src/derive_reflect/tuple_struct.rs
@@ -123,9 +123,11 @@ fn expand_reflect(
                 }
             });
 
+        let fields_len = fields.len();
+
         quote! {
             fn to_value(&self) -> Value {
-                let value = TupleStructValue::default();
+                let value = TupleStructValue::with_capacity(#fields_len);
                 #(#code_for_fields)*
                 value.into()
             }